requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.3.7
langchain==0.0.335
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
Pillow==10.1.0
python-dotenv==1.0.0
pydantic-settings==2.2.1
